Move core/notifier to pkg/notifier
Update package reference in related components
