+++ src/pkg/notification/hook/hook.go
@@ -10,10 +10,10 @@ import (
 	cModels "github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/common/utils/log"
 	"github.com/goharbor/harbor/src/core/config"
-	"github.com/goharbor/harbor/src/core/notifier/model"
 	"github.com/goharbor/harbor/src/core/utils"
 	"github.com/goharbor/harbor/src/pkg/notification/job"
 	"github.com/goharbor/harbor/src/pkg/notification/job/manager"
+	"github.com/goharbor/harbor/src/pkg/notifier/model"
 )
 
 // Manager send hook
