+++ src/pkg/notification/policy/manager/manager.go
@@ -10,8 +10,8 @@ import (
 	commonhttp "github.com/goharbor/harbor/src/common/http"
 	"github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/common/utils/log"
-	notifierModel "github.com/goharbor/harbor/src/core/notifier/model"
 	"github.com/goharbor/harbor/src/pkg/notification/model"
+	notifierModel "github.com/goharbor/harbor/src/pkg/notifier/model"
 )
 
 // DefaultManager ...
