+++ src/pkg/notifier/handler/notification/http_handler.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 
 	"github.com/goharbor/harbor/src/common/job/models"
-	"github.com/goharbor/harbor/src/core/notifier/model"
 	"github.com/goharbor/harbor/src/jobservice/job"
 	"github.com/goharbor/harbor/src/pkg/notification"
+	"github.com/goharbor/harbor/src/pkg/notifier/model"
 )
 
 // HTTPHandler preprocess http event data and start the hook processing
