+++ src/pkg/notifier/handler/notification/processor.go
@@ -9,9 +9,9 @@ import (
 	"github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/common/utils/log"
 	"github.com/goharbor/harbor/src/core/config"
-	"github.com/goharbor/harbor/src/core/notifier/event"
-	notifyModel "github.com/goharbor/harbor/src/core/notifier/model"
 	"github.com/goharbor/harbor/src/pkg/notification"
+	"github.com/goharbor/harbor/src/pkg/notifier/event"
+	notifyModel "github.com/goharbor/harbor/src/pkg/notifier/model"
 )
 
 // getNameFromImgRepoFullName gets image name from repo full name with format `repoName/imageName`
