+++ src/pkg/notifier/handler/notification/quota_handler.go
@@ -6,9 +6,9 @@ import (
 	"github.com/goharbor/harbor/src/common/models"
 	"github.com/goharbor/harbor/src/common/utils/log"
 	"github.com/goharbor/harbor/src/core/config"
-	"github.com/goharbor/harbor/src/core/notifier/model"
-	notifyModel "github.com/goharbor/harbor/src/core/notifier/model"
 	"github.com/goharbor/harbor/src/pkg/notification"
+	"github.com/goharbor/harbor/src/pkg/notifier/model"
+	notifyModel "github.com/goharbor/harbor/src/pkg/notifier/model"
 )
 
 // QuotaPreprocessHandler preprocess image event data
